fix: перенес классы в другой пакет
